changed some suggestions based on the codereview
Changed FEMS to Fenecon, (folder and Integrationplugin)
removed unnce. switch
Removed debian folder,
Moved Installation to debian main
Edit debian/control for Fenecon plugin
Added/edited Translation files
