--- fems/debian/changelog
@@ -1,5 +1,0 @@
-nymea-plugin-fems (0.0.1) xenial; urgency=medium
-
-  * Initial release.
-
- -- a <a@a.com>  Mon, 21 Aug 2017 10:12:00 +0000
