--- fems/debian/control
@@ -1,25 +1,0 @@
-Source: nymea-plugin-fems
-Section: utils
-Priority: options
-Maintainer: a <a@a.com>
-Build-depends: debhelper (>= 9.0.0),
-               libnymea1-dev,
-               nymea-dev-tools:native,
-               qt5-qmake:native,
-               pkg-config,
-Standards-Version: 3.9.3
-
-Package: nymea-plugin-fems
-Architecture: any
-Multi-Arch: same
-Section: libs
-Depends: ${shlibs:Depends},
-         ${misc:Depends},
-Description: nymea.io plugin for fems
- The nymea daemon is a plugin based IoT (Internet of Things) server. The
- server works like a translator for devices, things and services and
- allows them to interact.
- With the powerful rule engine you are able to connect any device available
- in the system and create individual scenes and behaviors for your environment.
- .
- This package will install the nymea.io plugin for fems
