--- fems/debian/copyright
@@ -1,14 +1,0 @@
-Format: http://www.debian.org/doc/packaging-manuals/copyright-format/1.0/
-Upstream-Name: ${ProjectName}
-Upstream-Contact: a <a@a.com>
-Copyright: 2020, a 
-
-
-License: LGPL-3
- On Debian systems, the complete text of the GNU Lesser General
- Public License can be found in `/usr/share/common-licenses/LGPL-3'.
-
-Files: *
-License: LGPL-3
-Copyright: a <a@a.com>
-
