--- fems/debian/nymea-plugin-fems.install.in
@@ -1,1 +1,0 @@
-usr/lib/@DEB_HOST_MULTIARCH@/nymea/plugins/libnymea_integrationpluginfems.so
