+++ fenecon/translations/19e10c9b-2dea-4aaf-bbfc-3c28b6450749-de.ts
@@ -1,27 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!DOCTYPE TS>
-<TS version="2.1">
+<TS version="2.1" language="de_DE">
 <context>
-    <name>IntegrationPluginFems</name>
+    <name>IntegrationPluginFenecon</name>
     <message>
         <location filename="../integrationpluginfems.cpp" line="58"/>
         <source>Please enter your login credentials for the FEMS connection</source>
-        <translation type="unfinished"></translation>
+        <translation type="unfinished">Bitte die Logindaten für die FEMS-Verbindung angeben.</translation>
     </message>
     <message>
         <location filename="../integrationpluginfems.cpp" line="127"/>
         <source>The device does not reply to our requests. Please verify that the FEMS API is enabled on the device.</source>
-        <translation type="unfinished"></translation>
+        <translation type="unfinished">Das FEMS reagiert nicht auf die Anfragen.Bitte sicherstellen, dass die FEMS Schnittstelle aktiviert ist.</translation>
     </message>
     <message>
         <location filename="../integrationpluginfems.cpp" line="132"/>
         <source>The device is not reachable.</source>
-        <translation type="unfinished"></translation>
+        <translation type="unfinished">Das Gerät ist nicht erreichbar.</translation>
     </message>
     <message>
         <location filename="../integrationpluginfems.cpp" line="145"/>
         <source>The data received from the device could not be processed because the format is unknown.</source>
-        <translation type="unfinished"></translation>
+        <translation type="unfinished">Die Daten konnten nicht verarbeitet werden. Format nicht unterstützt.</translation>
     </message>
 </context>
 </TS>
