+++ fenecon/translations/19e10c9b-2dea-4aaf-bbfc-3c28b6450749-en_US.ts
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="en_US">
+<context>
+    <name>IntegrationPluginFenecon</name>
+    <message>
+        <location filename="../integrationpluginfems.cpp" line="58"/>
+        <source>Please enter your login credentials for the FEMS connection</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../integrationpluginfems.cpp" line="127"/>
+        <source>The device does not reply to our requests. Please verify that the FEMS API is enabled on the device.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../integrationpluginfems.cpp" line="132"/>
+        <source>The device is not reachable.</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location filename="../integrationpluginfems.cpp" line="145"/>
+        <source>The data received from the device could not be processed because the format is unknown.</source>
+        <translation type="unfinished"></translation>
+    </message>
+</context>
+</TS>
