+++ nymea-plugins.pro
@@ -24,7 +24,7 @@ PLUGIN_DIRS = \
     espuino             \
     evbox               \
     fastcom             \
-    fems		        \
+    fenecon             \
     flowercare          \
     fronius             \
     garadget            \
